Clean up readlog tests. Fix indentation.
